doc: opportunity sorting
This surprisingly took a little longer than I thought to do since it
needed to change the structure of the variable passed to the
templates. The loop that prepares the list of opportunities now has a
little tidbit above it to explain how it's set up and what it does.